requirements.txt
arrow
discord
PyYAML
pillow
python-Levenshtein
fuzzywuzzy